* Created a reusable component that wraps the context-menu and prevents unwanted focus switching (PR #1914)
Fixes #1887
