+++ src/components/shared/CallNodeContextMenu.js
@@ -4,8 +4,8 @@
 
 // @flow
 import React, { PureComponent, Fragment } from 'react';
-import ReactDOM from 'react-dom';
-import { ContextMenu, MenuItem } from 'react-contextmenu';
+import { MenuItem } from 'react-contextmenu';
+import ContextMenu from '../shared/ContextMenu';
 import explicitConnect from '../../utils/connect';
 import { selectedThreadSelectors } from '../../selectors/per-thread';
 import { funcHasRecursiveCall } from '../../profile-logic/transforms';
@@ -68,11 +68,6 @@ type State = {|
 require('./CallNodeContextMenu.css');
 
 class CallNodeContextMenu extends PureComponent<Props, State> {
-  _contextMenu: ContextMenu | null = null;
-  _takeContextMenuRef = (contextMenu: ContextMenu | null) => {
-    this._contextMenu = contextMenu;
-  };
-
   constructor(props: Props) {
     super(props);
     this.state = {
@@ -358,25 +353,6 @@ class CallNodeContextMenu extends PureComponent<Props, State> {
     }
   }
 
-  _mouseDownHandler(event: MouseEvent): void {
-    event.preventDefault();
-  }
-
-  componentDidUpdate() {
-    if (this.state.isShown && this._contextMenu) {
-      // The context menu component does not expose a reference to its internal
-      // DOM node so using findDOMNode is currently unavoidable.
-      // eslint-disable-next-line react/no-find-dom-node
-      const contextMenuNode = ReactDOM.findDOMNode(this._contextMenu);
-      if (contextMenuNode) {
-        // There's no need to remove this event listener since the component is
-        // never unmounted. Duplicate event listeners will also be discarded
-        // automatically so we don't need to handle that.
-        contextMenuNode.addEventListener('mousedown', this._mouseDownHandler);
-      }
-    }
-  }
-
   renderContextMenuContents() {
     const {
       selectedCallNodeIndex,
@@ -494,16 +470,10 @@ class CallNodeContextMenu extends PureComponent<Props, State> {
     return (
       <ContextMenu
         id="CallNodeContextMenu"
-        ref={this._takeContextMenuRef}
         onShow={this._showMenu}
         onHide={this._hideMenu}
       >
-        {this.state.isShown ? (
-          this.renderContextMenuContents()
-        ) : (
-          // ContextMenu expects at least 1 child.
-          <div />
-        )}
+        {this.state.isShown ? this.renderContextMenuContents() : null}
       </ContextMenu>
     );
   }
